requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asciinema==2.0.2
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
btrfsutil==5.13.1
CacheControl==0.12.6
cached-property==1.5.2
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.7
discord.py==1.7.3
distlib==0.3.2
distro==1.5.0
dnspython==1.16.0
docker==5.0.0
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
html5lib==1.1
idna==3.2
importlib-metadata==4.6.1
iotop==0.6
jsonschema==3.2.0
lensfun==0.3.95
Markdown==3.3.4
meld==3.20.4
more-itertools==8.8.0
msgpack==1.0.2
multidict==5.1.0
openshot-qt==2.5.1
ordered-set==4.0.2
packaging==20.9
paramiko==2.7.2
pep517==0.11.0
pikaur==1.7
ply==3.11
progress==1.5
psutil==5.8.0
pyalpm==0.10.6
pycairo==1.20.0
pycparser==2.20
PyGObject==3.40.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dotenv==0.19.0
PyYAML==5.4.1
pyzmq==22.2.1
ranger-fm==1.9.3
Reflector==2021.7.8.19.52.52
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
six==1.16.0
terminator==2.1.1
texttable==1.6.4
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
validate==5.1.0.dev0
webencodings==0.5.1
websocket-client==0.59.0
yarl==1.6.3
zipp==3.5.0
